+++ CHANGES.rst
@@ -5,8 +5,14 @@ Changelog
 4.2.9rc3 (unreleased)
 ---------------------
 
-- Nothing changed yet.
+- Fixed custom advice WF UI:
 
+  - do not display `given_by` information when using the default advice workflow;
+  - in `onAdviceTransition` event, notify `AdviceAfterTransitionEvent` after
+    `hidden_during_redaction` auto set to `False` management;
+  - adapted import_data to manage `advisersConfig`.
+
+  [gbastien]
 
 4.2.9rc2 (2024-02-26)
 ---------------------